requirements.txt
# файл шоб швидко скачати і розвернути на сервері но хуй я його напишу і буду як додік в терміналі це все вводити
aiogram
apscheduler
selenium
webdriver-manager
requests
opencv-python-headless
numpy
pillow
aiosqlite
pytz